Revert "Use Long integers for OCLC work ids. Some are too large for ints."
This reverts commit 9e8feccc78b0dffa8beed96829429fbf1ebf3d35.

+++ rdfToSolr/src/main/java/edu/cornell/library/integration/indexer/documentPostProcess/UpdateSolrInventoryDB.java
@@ -278,10 +278,10 @@ public class UpdateSolrInventoryDB implements DocumentPostProcess{
 			pstmt.setInt(1, oclcId);
 			ResultSet rs = pstmt.executeQuery();
 			while (rs.next()) {
-				long workid = rs.getLong(1);
+				int workid = rs.getInt(1);
 				insertStmt.setInt(1, bibid);
 				insertStmt.setInt(2, oclcId);
-				insertStmt.setLong(3, workid);
+				insertStmt.setInt(3, workid);
 				insertStmt.addBatch();
 			}
 		}
